Reduce execution almost done

+++ Language/HostLanguage.fs
@@ -12,7 +12,7 @@ module HostLanguage =
     let notused b =
         b
 
-    let worksize(comp, globalSize:int array, localSize:int array) =
+    let worksize(comp, globalSize:int64 array, localSize:int64 array) =
         comp
